feature: start storing jwt token inside cookie
